refactor: add a .gitignore file in k3d to ignore registry
This is more of a placeholder, but there needs to be a way of excluding
common files if required. The k3d providers don't have a registry folder,
so a .gitignore might be the way of achieving it. This would need to be
merged in with any other .gitignore folders, so we might need a better way
of achieving this.